+++ src/main/claude-usage/scanner.test.ts
@@ -1,10 +1,7 @@
 import { describe, expect, it } from 'vitest'
-import {
-  aggregateClaudeUsage,
-  attributeClaudeUsageTurns,
-  parseClaudeUsageFile,
-  parseClaudeUsageRecord
-} from './scanner'
+import { parseClaudeUsageFile, parseClaudeUsageRecord } from './transcript-record-parser'
+import { attributeClaudeUsageTurns } from './worktree-attribution'
+import { aggregateClaudeUsage } from './usage-aggregation'
 import { mkdtemp, rm, writeFile } from 'node:fs/promises'
 import { join } from 'node:path'
 import { tmpdir } from 'node:os'
